+++ crates/nostr-zapper/src/error.rs
@@ -10,17 +10,11 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum ZapperError {
     /// An error happened in the underlying zapper backend.
-    #[error("zapper: {0}")]
+    #[error(transparent)]
     Backend(Box<dyn std::error::Error + Send + Sync>),
-    /// Nostr error
-    #[error("nostr: {0}")]
-    Nostr(Box<dyn std::error::Error + Send + Sync>),
     /// Not supported
-    #[error("method not supported by current backend")]
+    #[error("not supported by current backend")]
     NotSupported,
-    /// Feature disabled
-    #[error("feature disabled for current backend")]
-    FeatureDisabled,
 }
 
 impl ZapperError {
@@ -33,16 +27,5 @@ impl ZapperError {
         E: std::error::Error + Send + Sync + 'static,
     {
         Self::Backend(Box::new(error))
-    }
-
-    /// Create a new `Nostr` error.
-    ///
-    /// Shorthand for `Error::Nostr(Box::new(error))`.
-    #[inline]
-    pub fn nostr<E>(error: E) -> Self
-    where
-        E: std::error::Error + Send + Sync + 'static,
-    {
-        Self::Nostr(Box::new(error))
     }
 }