+++ crates/nostr-zapper/src/lib.rs
@@ -8,7 +8,6 @@
 #![warn(rustdoc::bare_urls)]
 #![allow(unknown_lints)]
 
-use core::fmt;
 use std::sync::Arc;
 
 pub extern crate nostr;
@@ -33,7 +32,7 @@ pub enum ZapperBackend {
 }
 
 /// A type-erased [`NostrZapper`].
-pub type DynNostrZapper = dyn NostrZapper<Err = ZapperError>;
+pub type DynNostrZapper = dyn NostrZapper;
 
 /// A type that can be type-erased into `Arc<dyn NostrZapper>`.
 pub trait IntoNostrZapper {
@@ -52,22 +51,16 @@ where
     T: NostrZapper + Sized + 'static,
 {
     fn into_nostr_zapper(self) -> Arc<DynNostrZapper> {
-        Arc::new(EraseNostrZapperError(self))
+        Arc::new(self)
     }
 }
 
-// Turns a given `Arc<T>` into `Arc<DynNostrZapper>` by attaching the
-// NostrZapper impl vtable of `EraseNostrZapperError<T>`.
 impl<T> IntoNostrZapper for Arc<T>
 where
     T: NostrZapper + 'static,
 {
     fn into_nostr_zapper(self) -> Arc<DynNostrZapper> {
-        let ptr: *const T = Arc::into_raw(self);
-        let ptr_erased = ptr as *const EraseNostrZapperError<T>;
-        // SAFETY: EraseNostrZapperError is repr(transparent) so T and
-        //         EraseNostrZapperError<T> have the same layout and ABI
-        unsafe { Arc::from_raw(ptr_erased) }
+        self
     }
 }
 
@@ -75,37 +68,11 @@ where
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 pub trait NostrZapper: AsyncTraitDeps {
-    /// Error
-    type Err: From<ZapperError> + Into<ZapperError>;
-
     /// Name of the backend zapper used (ex. WebLN, NWC, ...)
     fn backend(&self) -> ZapperBackend;
 
     /// Pay invoice
-    async fn pay(&self, invoice: String) -> Result<(), Self::Err>;
-}
-
-#[repr(transparent)]
-struct EraseNostrZapperError<T>(T);
-
-impl<T: fmt::Debug> fmt::Debug for EraseNostrZapperError<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-impl<T: NostrZapper> NostrZapper for EraseNostrZapperError<T> {
-    type Err = ZapperError;
-
-    fn backend(&self) -> ZapperBackend {
-        self.0.backend()
-    }
-
-    async fn pay(&self, invoice: String) -> Result<(), Self::Err> {
-        self.0.pay(invoice).await.map_err(Into::into)
-    }
+    async fn pay(&self, invoice: String) -> Result<(), ZapperError>;
 }
 
 /// Alias for `Send` on non-wasm, empty trait (implemented by everything) on